Use TimerMilli/TimerMicro instead of TimerA/TimerB
Because some small MSP430 MCU have no TimerB, we initialize TimerA for 32kHz
timer instead of 1MiHz. If a MCU has more than 2 timers, then we initialize
the second timer for 1MiHz, that may be TimerB or TimerA1.

Note that upstream tinyos-prod sets up TimerA for 1MiHz timer and TimerB for
32kiHz timer.